pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "cloudvirt"
authors = [
    {name = "gottaeat", email = "mss@tutanota.de"}
]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"
]
dynamic = [
    "version",
    "description"
]

dependencies = [
    "pycdlib >= 1.14.0",
    "PyYAML >= 6.0.0",
    "passlib >= 1.6.0",
    "libvirt-python >= 9.5.0",
]

[project.urls]
Home = "https://github.com/gottaeat/cloudvirt"

[project.scripts]
cloudvirt = "cloudvirt.cli:run"
